backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
pandas==2.2.2
python-multipart==0.0.9
groq>=0.31.0
python-dotenv==1.0.1
feedparser>=6.0.10